Fix "maximum update depth exceeded" error for long verses.
In order to draw link lines correctly between text segments,
we keep a `ref` for every textSegment in react state (using setState).
React has a limit on how many times `setState` can be called in a nested
fashion so that infinite loops don't occur during a render. The maximum
update depth is unfortunately to small of a number to set each textSegment
ref in state as they are rendered.

To fix this, we store a map of the textSegment refs within the component closure.
As textSegment refs are created they are accumulated in the map. Once the map reaches
the size of the source or target segment array, then the react state is updated.

This greatly reduces the number of times `setState` is called despite the size of
the source or target segment array. We will learn in time if this "hack" will cause
other problems. For now it seems stable.