Simple bridge/source flip in line view.
This was a spike to uncover the issues involved if we wanted to
show the source text as the center layer in the line view, with
bridge/reference above (linked to source) and target below
(also linked to source). The biggest hurdle is that this actually
entails a change in definition of a user link when in bridge mode.

Need to think on this some more.